requirements.txt
langchain-core~=1.2.7
langchain-google-genai~=4.2.0
langchain-community~=0.4.1
langgraph~=1.0.6
python-dotenv~=1.2.1